Changed prediction test output image to a 32-bit float

+++ tests/test_core.py
@@ -123,4 +123,4 @@ class TestTrainer(unittest.TestCase):
 
         tif.imsave(os.path.join(IMAGE_PATH,
                                 "test_prediction.tif"),
-                   output_volume[300].getArray())
+                   output_volume[300].getArray().astype(np.float32))
